\name{hopskip-package}
\alias{hopskip-package}
\alias{hopskip}
\docType{package}
\title{
Simulates spatial spread of insects in a city.
}
\description{
This simulation is a discrete stochastic process in continuous time
on a spatial domain of houses in a city with streets separating blocks.
Houses become infected and infection spreads to other houses with a kernel.
}
\details{
\tabular{ll}{
Package: \tab hopskip\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-23\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Drew Dolgert

Maintainer: Andrew Dolgert <adolgert@cornell.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
